+++ CHANGELOG.md
@@ -2,16 +2,38 @@
 
 This project roughly adheres to [Semantic Versioning](http://semver.org/). For 0.x.y releases, `x` is the major version in semver, while `y` is the minor version.
 
-## 0.8.0 - 2019-06-13
+## ctp-common 0.9.0, ctp-md 0.10.0, ctp-trader 0.10.0 - 2020-05-13
+
+* Fix memory leaks
+* Fix unsafe function signature
+* Fix clippy warnings
+
+## ctp-md 0.9.0, ctp-trader 0.9.0 - 2020-01-23
+
+* Upgrade `crossbeam-channel` to 0.4.0
+
+## ctp-common 0.8.2 - 2020-01-20
+
+* Fix clippy warnings
+
+## ctp-common 0.8.1, ctp-md 0.8.1, ctp-trader 0.8.2 - 2020-01-12
+
+* Upgrade to 2018 edition
+
+## ctp-trader 0.8.1 - 2019-06-13
+
+* Fix `crossbeam-channel` version
+
+## ctp-common 0.8.0, ctp-md 0.8.0, ctp-trader 0.8.0 - 2019-06-13
 
 * Support transparent regulatory API for prod environment (api-ctp 6.3.15)
 
-## 0.7.0 - 2019-06-13
+## ctp-common 0.7.0, ctp-md 0.8.0, ctp-trader 0.8.0 - 2019-06-13
 
 * Support transparent regulatory API for test environment (api-ctp 6.3.13)
 * Upgrade `crossbeam-channel` to 0.3.8
 * Upgrade `simple-error` to 0.2.0
 
-## 0.6.0 - 2018-09-03
+## ctp-common 0.6.0, ctp-md 0.8.0, ctp-trader 0.8.0 - 2018-09-03
 
 * Support `crossbeam-channel 0.2.4` through feature `channel`